requirements.txt
openai>=1.30.0
beautifulsoup4>=4.12.3
flask>=3.0.0
gunicorn>=22.0.0
lxml>=5.2.2
pydantic>=2.7.4
pydantic-settings>=2.3.4
python-dotenv>=1.0.1
python-docx>=1.1.2
pypdf>=4.3.1
PyYAML>=6.0.2
readability-lxml>=0.8.1
reportlab>=4.2.2
requests>=2.32.3
tenacity>=8.5.0
